requirements.txt
tiktoken
chromadb
langchain
textract
llama-cpp-python
pypdf
python-dotenv
